move update model schemas to model definition
