+++ sqlmesh/core/loader.py
@@ -7,12 +7,10 @@ import logging
 import os
 import typing as t
 from collections import defaultdict
-from concurrent.futures import as_completed
 from dataclasses import dataclass
 from pathlib import Path
 
-from sqlglot.errors import SchemaError, SqlglotError
-from sqlglot.schema import MappingSchema
+from sqlglot.errors import SqlglotError
 
 from sqlmesh.core import constants as c
 from sqlmesh.core.audit import Audit, ModelAudit, StandaloneAudit, load_multiple_audits
@@ -23,16 +21,13 @@ from sqlmesh.core.model import (
     Model,
     ExternalModel,
     ModelCache,
-    OptimizedQueryCache,
     SeedModel,
     create_external_model,
     load_sql_based_model,
 )
-from sqlmesh.core.model.cache import load_optimized_query_and_mapping, optimized_query_cache_pool
 from sqlmesh.core.model import model as model_registry
 from sqlmesh.core.signal import signal
 from sqlmesh.utils import UniqueKeyDict, sys_path
-from sqlmesh.utils.dag import DAG
 from sqlmesh.utils.errors import ConfigError
 from sqlmesh.utils.jinja import JinjaMacroRegistry, MacroExtractor
 from sqlmesh.utils.metaprogramming import import_python_file
@@ -622,100 +617,3 @@ class SqlMeshLoader(Loader):
                     self._loader.context.gateway or self._loader.config.default_gateway_name,
                 ]
             )
-
-
-# TODO: consider moving this to context
-def update_model_schemas(
-    dag: DAG[str],
-    models: UniqueKeyDict[str, Model],
-    config_path: Path,
-) -> None:
-    schema = MappingSchema(normalize=False)
-    optimized_query_cache: OptimizedQueryCache = OptimizedQueryCache(config_path / c.CACHE)
-
-    if c.MAX_FORK_WORKERS == 1:
-        _update_model_schemas_sequential(dag, models, schema, optimized_query_cache)
-    else:
-        _update_model_schemas_parallel(dag, models, schema, optimized_query_cache)
-
-
-def _update_schema_with_model(schema: MappingSchema, model: Model) -> None:
-    columns_to_types = model.columns_to_types
-    if columns_to_types:
-        try:
-            schema.add_table(model.fqn, columns_to_types, dialect=model.dialect)
-        except SchemaError as e:
-            if "nesting level:" in str(e):
-                logger.error(
-                    "SQLMesh requires all model names and references to have the same level of nesting."
-                )
-            raise
-
-
-def _update_model_schemas_sequential(
-    dag: DAG[str],
-    models: UniqueKeyDict[str, Model],
-    schema: MappingSchema,
-    optimized_query_cache: OptimizedQueryCache,
-) -> None:
-    for name in dag.sorted:
-        model = models.get(name)
-
-        # External models don't exist in the context, so we need to skip them
-        if not model:
-            continue
-
-        model.update_schema(schema)
-        optimized_query_cache.with_optimized_query(model)
-        _update_schema_with_model(schema, model)
-
-
-def _update_model_schemas_parallel(
-    dag: DAG[str],
-    models: UniqueKeyDict[str, Model],
-    schema: MappingSchema,
-    optimized_query_cache: OptimizedQueryCache,
-) -> None:
-    futures = set()
-    graph = {
-        model: {dep for dep in deps if dep in models}
-        for model, deps in dag._dag.items()
-        if model in models
-    }
-
-    def process_models(completed_model: t.Optional[Model] = None) -> None:
-        for name in list(graph):
-            deps = graph[name]
-
-            if completed_model:
-                deps.discard(completed_model.fqn)
-
-            if not deps:
-                del graph[name]
-                model = models[name]
-                futures.add(
-                    executor.submit(
-                        load_optimized_query_and_mapping,
-                        model,
-                        mapping={
-                            parent: models[parent].columns_to_types
-                            for parent in model.depends_on
-                            if parent in models
-                        },
-                    )
-                )
-
-    with optimized_query_cache_pool(optimized_query_cache) as executor:
-        process_models()
-
-        while futures:
-            for future in as_completed(futures):
-                futures.remove(future)
-                fqn, entry_name, data_hash, metadata_hash, mapping_schema = future.result()
-                model = models[fqn]
-                model._data_hash = data_hash
-                model._metadata_hash = metadata_hash
-                model.set_mapping_schema(mapping_schema)
-                optimized_query_cache.with_optimized_query(model, entry_name)
-                _update_schema_with_model(schema, model)
-                process_models(completed_model=model)
