+++ sqlmesh/core/model/__init__.py
@@ -37,4 +37,5 @@ from sqlmesh.core.model.kind import (
     model_kind_validator as model_kind_validator,
 )
 from sqlmesh.core.model.meta import ModelMeta as ModelMeta
+from sqlmesh.core.model.schema import update_model_schemas as update_model_schemas
 from sqlmesh.core.model.seed import Seed as Seed
